Use a session
The benefit is that the http_call implementation is a lot smaller
because it can rely on the session. Another is that it's more efficient
due to connection reuse.

Another small improvement is using json instead of data which
automatically sets the Content-Type header to application/json. That
does mean we need requests 2.4.2 or higher.